Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/iframe-inside-csp-expected.txt

ALERT: PASS (1/2): Script can execute
ALERT: PASS (2/2): Eval works
